Style
Reworked how the images on the cardviews are sized. Changed mismatching accent colors in the two style files. Reworked how images are sized in fragment_article_detail. Deleted holder.xml that was acting as a standin for fragment_article_detail while I was working on making changes to the latter file. Fragment_article_detail is once again being used instead of holder.xml